+++ list.go
@@ -3,11 +3,14 @@ package main
 import (
 	"fmt"
 	"go/build"
+	"go/parser"
+	"go/token"
 	"log"
 	"os"
 	"path/filepath"
-	"sort"
+	"strconv"
 	"strings"
+	"unicode"
 
 	pathpkg "path"
 )
@@ -16,10 +19,6 @@ var (
 	buildContext = build.Default
 	gorootSrc    = filepath.Join(buildContext.GOROOT, "src")
 )
-
-func init() {
-	buildContext.UseAllFiles = true
-}
 
 // packageContext is used to track an import and which package imported it.
 type packageContext struct {
@@ -51,7 +50,7 @@ func (ds *depScanner) Continue() bool {
 
 // Add a package and imports to the depScanner. Skips already processed package/import combos
 func (ds *depScanner) Add(pkg *build.Package, imports ...string) {
-Next:
+NextImport:
 	for _, i := range imports {
 		if i == "C" {
 			i = "runtime/cgo"
@@ -60,17 +59,17 @@ Next:
 		for _, epc := range ds.processed {
 			if epc == pc {
 				debugln("ctxts epc == pc, skipping", epc, pc)
-				continue Next
+				continue NextImport
 			}
 			if pc.pkg.Dir == epc.pkg.Dir && pc.imp == epc.imp {
 				debugln("ctxts epc.pkg.Dir == pc.pkg.Dir && pc.imp == epc.imp, skipping", epc.pkg.Dir, pc.imp)
-				continue Next
+				continue NextImport
 			}
 		}
 		for _, epc := range ds.todo {
 			if epc == pc {
 				debugln("ctxts epc == pc, skipping", epc, pc)
-				continue Next
+				continue NextImport
 			}
 		}
 		debugln("Adding pc:", pc)
@@ -92,8 +91,6 @@ func listPackage(path string) (*Package, error) {
 	var dir string
 	var lp *build.Package
 	var err error
-	deps := make(map[string]bool)
-	imports := make(map[string]bool)
 	if build.IsLocalImport(path) {
 		dir = path
 		if !filepath.IsAbs(dir) {
@@ -102,15 +99,19 @@ func listPackage(path string) (*Package, error) {
 				dir = abs
 			}
 		}
-		lp, err = buildContext.ImportDir(dir, 0)
+		lp, err = buildContext.ImportDir(dir, build.FindOnly)
 	} else {
 		dir, err = os.Getwd()
 		if err != nil {
 			return nil, err
 		}
-		lp, err = buildContext.Import(path, dir, 0)
-		lp, err = checkGoroot(lp, err)
+		lp, err = buildContext.Import(path, dir, build.FindOnly)
+		if lp.Goroot {
+			// If it's in the GOROOT, just import it
+			lp, err = buildContext.Import(path, dir, 0)
+		}
 	}
+	fillPackage(lp)
 	p := &Package{
 		Dir:            lp.Dir,
 		Root:           lp.Root,
@@ -142,7 +143,7 @@ func listPackage(path string) (*Package, error) {
 			for base := ip.Dir; base != ip.Root; base = filepath.Dir(base) {
 				dir := filepath.Join(base, "vendor", i)
 				debugln("dir:", dir)
-				dp, err = buildContext.ImportDir(dir, 0)
+				dp, err = buildContext.ImportDir(dir, build.FindOnly)
 				if err != nil {
 					if os.IsNotExist(err) {
 						continue
@@ -155,25 +156,17 @@ func listPackage(path string) (*Package, error) {
 			}
 		}
 		// Wasn't found above, so resolve it using the build.Context
-		dp, err = buildContext.Import(i, ip.Dir, 0)
+		dp, err = buildContext.Import(i, ip.Dir, build.FindOnly)
+		if dp.Goroot {
+			// If it's in the Goroot, just import the package
+			dp, err = buildContext.Import(i, ip.Dir, 0)
+		}
 		if err != nil {
-			if dp.Goroot {
-				// If it's in the GOROOT we can probably recover
-				switch err.(type) {
-				case *build.MultiplePackageError:
-					debugln("MultiplePackageError, importing Goroot package, trying w/o UseAllFiles")
-					dp, err = checkGoroot(dp, err)
-				default:
-					ppln(err)
-					debugln(err.Error())
-					panic("Unknown error importing GOROOT package: " + dp.ImportPath)
-				}
-			} else {
-				debugln("Warning: Error importing dependent package")
-				ppln(err)
-			}
+			debugln("Warning: Error importing dependent package")
+			ppln(err)
 		}
 	Found:
+		fillPackage(dp)
 		ppln(dp)
 		if dp.Goroot {
 			// Treat packages discovered to be in the GOROOT as if the package we're looking for is importing them
@@ -185,21 +178,84 @@ func listPackage(path string) (*Package, error) {
 		debugln("ip:", ip)
 		if lp == ip {
 			debugln("lp == ip")
-			imports[dp.ImportPath] = true
+			p.Imports = append(p.Imports, dp.ImportPath)
 		}
-		deps[dp.ImportPath] = true
+		p.Deps = append(p.Deps, dp.ImportPath)
 	}
-	for k := range deps {
-		p.Deps = append(p.Deps, k)
-	}
-	for k := range imports {
-		p.Imports = append(p.Imports, k)
-	}
-	sort.Strings(p.Imports)
-	sort.Strings(p.Deps)
+	p.Imports = uniq(p.Imports)
+	p.Deps = uniq(p.Deps)
 	debugln("Looking For Package:", path, "in", dir)
 	ppln(p)
 	return p, nil
+}
+
+// fillPackage full of info. Assumes a build.Package discovered in build.FindOnly mode
+func fillPackage(p *build.Package) error {
+
+	if p.Goroot {
+		return nil
+	}
+
+	var buildMatch = "+build "
+	var buildFieldSplit = func(r rune) bool {
+		return unicode.IsSpace(r) || r == ','
+	}
+
+	debugln("Filling package:", p.ImportPath, "from", p.Dir)
+	gofiles, err := filepath.Glob(filepath.Join(p.Dir, "*.go"))
+	if err != nil {
+		debugln("Error globbing", err)
+		return err
+	}
+
+	var testImports []string
+	var imports []string
+NextFile:
+	for _, file := range gofiles {
+		debugln(file)
+		fset := token.NewFileSet()
+		pf, err := parser.ParseFile(fset, file, nil, parser.ParseComments)
+		if err != nil {
+			return err
+		}
+		testFile := strings.HasSuffix(file, "_test.go")
+		fname := filepath.Base(file)
+		if testFile {
+			p.TestGoFiles = append(p.TestGoFiles, fname)
+		} else {
+			p.GoFiles = append(p.GoFiles, fname)
+		}
+		if len(pf.Comments) > 0 {
+			for _, c := range pf.Comments {
+				ct := c.Text()
+				if i := strings.Index(ct, buildMatch); i != -1 {
+					for _, b := range strings.FieldsFunc(ct[i+len(buildMatch):], buildFieldSplit) {
+						//TODO: appengine is a special case for now: https://github.com/tools/godep/issues/353
+						if b == "ignore" || b == "appengine" {
+							p.IgnoredGoFiles = append(p.IgnoredGoFiles, fname)
+							continue NextFile
+						}
+					}
+				}
+			}
+		}
+		for _, is := range pf.Imports {
+			name, err := strconv.Unquote(is.Path.Value)
+			if err != nil {
+				return err // can't happen?
+			}
+			if testFile {
+				testImports = append(testImports, name)
+			} else {
+				imports = append(imports, name)
+			}
+		}
+	}
+	imports = uniq(imports)
+	testImports = uniq(testImports)
+	p.Imports = imports
+	p.TestImports = testImports
+	return nil
 }
 
 // All of the following functions were vendored from go proper. Locations are noted in comments, but may change in future Go versions.